# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FACTA — Independent product intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all goals
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalSummary"
  /goals/{slug}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get goal detail with meal contexts
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Goal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user-goals/{sessionId}:
    get:
      operationId: getUserGoals
      tags: [goals]
      summary: Get user's active goals
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: User goals with profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGoalsResponse"
    put:
      operationId: setUserGoals
      tags: [goals]
      summary: Set user goals (max 2 active)
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserGoalsInput"
      responses:
        "200":
          description: Updated user goals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGoalsResponse"
  /profile/{sessionId}:
    get:
      operationId: getUserProfile
      tags: [goals]
      summary: Get user onboarding profile
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: saveUserProfile
      tags: [goals]
      summary: Save / update user onboarding profile
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  # ── Goal Fit ───────────────────────────────────────────────────────────────
  /goal-fit/{productId}/{goalSlug}:
    get:
      operationId: getGoalFit
      tags: [goals]
      summary: Compute Goal Fit for a product × goal pair
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
        - name: goalSlug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Goal fit result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalFitResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Meal Logs ─────────────────────────────────────────────────────────────
...
  /admin/goals:
    post:
      operationId: adminCreateGoal
      tags: [admin]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSummary"
  /admin/goals/{id}:
    patch:
      operationId: adminUpdateGoal
      tags: [admin]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGoalInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalSummary"
  /admin/goal-rulesets:
    post:
      operationId: adminPublishGoalRuleset
      tags: [admin]
      summary: Publish a goal ruleset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGoalRulesetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  goalId: { type: integer }
                  version: { type: string }
                  status: { type: string }
  /admin/meal-contexts/{id}:
    patch:
      operationId: adminUpdateMealContext
      tags: [admin]
      summary: Update meal context guidance
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMealContextInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: integer }
                  status: { type: string }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    GoalSummary:
      type: object
      required: [id, slug, name, nameZh, status]
      properties:
        id:            { type: integer }
        slug:          { type: string }
        name:          { type: string }
        nameZh:        { type: string }
        description:   { type: ["string", "null"] }
        descriptionZh: { type: ["string", "null"] }
        status:        { type: string }
        icon:          { type: ["string", "null"] }
        sortOrder:     { type: integer }
    MealContextSummary:
      type: object
      required: [id, meal, headlineZh, chooseMoreZh, chooseLessZh]
      properties:
        id:            { type: integer }
        meal:          { type: string }
        headline:      { type: string }
        headlineZh:    { type: string }
        chooseMore:    { type: array, items: { type: string } }
        chooseMoreZh:  { type: array, items: { type: string } }
        chooseLess:    { type: array, items: { type: string } }
        chooseLessZh:  { type: array, items: { type: string } }
        ctaText:       { type: ["string", "null"] }
        ctaTextZh:     { type: ["string", "null"] }
    GoalDetail:
      allOf:
        - $ref: "#/components/schemas/GoalSummary"
        - type: object
          properties:
            mealContexts:
              type: array
              items:
                $ref: "#/components/schemas/MealContextSummary"
            guides:
              type: array
              items:
                $ref: "#/components/schemas/GuideSummary"
    UserGoalsResponse:
      type: object
      required: [sessionId, activeGoals]
      properties:
        sessionId:    { type: string }
        activeGoals:
          type: array
          items:
            type: object
            properties:
              goalId:   { type: integer }
              goalSlug: { type: string }
              goalName: { type: string }
              goalNameZh: { type: string }
              priority: { type: string }
              status:   { type: string }
        profile:
          $ref: "#/components/schemas/UserProfile"
        onboardingCompleted: { type: boolean }
    UserGoalsInput:
      type: object
      required: [goals]
      properties:
        goals:
          type: array
          maxItems: 2
          items:
            type: object
            required: [goalId, priority]
            properties:
              goalId:   { type: integer }
              priority: { type: string }
    UserProfile:
      type: object
      required: [sessionId]
      properties:
        sessionId:          { type: string }
        preferredRetailers: { type: array, items: { type: string } }
        budgetTier:         { type: ["string", "null"] }
        wantsMealTiming:    { type: boolean }
        onboardingCompleted: { type: boolean }
        updatedAt:          { type: string }
    UserProfileInput:
      type: object
      properties:
        preferredRetailers: { type: array, items: { type: string } }
        budgetTier:         { type: ["string", "null"] }
        wantsMealTiming:    { type: boolean }
        onboardingCompletedAt: { type: ["string", "null"] }

    # ── Goal Fit ──────────────────────────────────────────────────────────
    FitReason:
      type: object
      required: [label, labelZh, positive]
      properties:
        label:    { type: string }
        labelZh:  { type: string }
        positive: { type: boolean }
    GoalFitResult:
      type: object
      required: [productId, goalSlug, fitLevel, goalRulesetVersion]
      properties:
        productId:           { type: integer }
        goalSlug:            { type: string }
        goalName:            { type: ["string", "null"] }
        goalNameZh:          { type: ["string", "null"] }
        fitLevel:            { type: string }
        fitReasons:
          type: array
          items:
            $ref: "#/components/schemas/FitReason"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/FitReason"
        breakfastFit:        { type: string }
        lunchFit:            { type: string }
        dinnerFit:           { type: string }
        snackFit:            { type: string }
        inputDataCompleteness: { type: ["number", "null"] }
        goalRulesetVersion:  { type: string }
        evaluatedAt:         { type: string }

    # ── Meal Logs ─────────────────────────────────────────────────────────
...
    GuideSummary:
      type: object
      required: [id, slug, titleZh, status]
      properties:
        id:          { type: integer }
        slug:        { type: string }
        title:       { type: string }
        titleZh:     { type: string }
        summaryZh:   { type: ["string", "null"] }
        summary:     { type: ["string", "null"] }
        goalId:      { type: ["integer", "null"] }
        coverImageUrl: { type: ["string", "null"] }
        status:      { type: string }
        publishedAt: { type: ["string", "null"] }
...
    AdminGoalInput:
      type: object
      properties:
        slug:          { type: string }
        name:          { type: string }
        nameZh:        { type: string }
        description:   { type: ["string", "null"] }
        descriptionZh: { type: ["string", "null"] }
        status:        { type: string }
        icon:          { type: ["string", "null"] }
        sortOrder:     { type: integer }
    AdminGoalRulesetInput:
      type: object
      required: [goalId, version, rules]
      properties:
        goalId:       { type: integer }
        version:      { type: string }
        description:  { type: ["string", "null"] }
        rules:        { type: object, additionalProperties: true }
        status:       { type: string }
    AdminMealContextInput:
      type: object
      properties:
        headline:      { type: string }
        headlineZh:    { type: string }
        chooseMore:    { type: array, items: { type: string } }
        chooseMoreZh:  { type: array, items: { type: string } }
        chooseLess:    { type: array, items: { type: string } }
        chooseLessZh:  { type: array, items: { type: string } }
        ctaText:       { type: ["string", "null"] }
        ctaTextZh:     { type: ["string", "null"] }
        status:        { type: string }
...