# lib/api-zod/src/generated/types/fitReason.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FACTA — Independent product intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface FitReason {
  label: string;
  labelZh: string;
  positive: boolean;
}