# lib/api-zod/src/generated/types/goalDetail.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FACTA — Independent product intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { GoalSummary } from './goalSummary';
import type { GuideSummary } from './guideSummary';
import type { MealContextSummary } from './mealContextSummary';

export type GoalDetail = GoalSummary & {
  mealContexts?: MealContextSummary[];
  guides?: GuideSummary[];
};