# lib/api-zod/src/generated/types/goalFitResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FACTA — Independent product intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { FitReason } from './fitReason';

export interface GoalFitResult {
  productId: number;
  goalSlug: string;
  /** @nullable */
  goalName?: string | null;
  /** @nullable */
  goalNameZh?: string | null;
  fitLevel: string;
  fitReasons?: FitReason[];
  warnings?: FitReason[];
  breakfastFit?: string;
  lunchFit?: string;
  dinnerFit?: string;
  snackFit?: string;
  /** @nullable */
  inputDataCompleteness?: number | null;
  goalRulesetVersion: string;
  evaluatedAt?: string;
}