# lib/api-zod/src/generated/types/goalSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FACTA — Independent product intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface GoalSummary {
  id: number;
  slug: string;
  name: string;
  nameZh: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionZh?: string | null;
  status: string;
  /** @nullable */
  icon?: string | null;
  sortOrder?: number;
}